move changes, add some more speedups

+++ test/Microsoft.TestPlatform.TestUtilities/IntegrationTestBuild.cs
@@ -53,7 +53,7 @@ public class IntegrationTestBuild : IntegrationTestBase
                 Debug.WriteLine("Starting to build.");
                 var sw = Stopwatch.StartNew();
 
-                var nugetCache = Path.GetFullPath(Path.Combine(Root, ".packages"));
+                var nugetCache = IntegrationTestEnvironment.TestAssetsNuGetCacheDirectory;
                 var packagesAreNew = UnzipExecutablePackages();
                 if (packagesAreNew)
                 {
@@ -486,13 +486,16 @@ public class IntegrationTestBuild : IntegrationTestBase
         }
 
         // Then clean all -dev and -ci packages from the cache to force updating from local source.
-        foreach (var packageDir in Directory.GetDirectories(nugetCache))
+        if (Directory.Exists(nugetCache))
         {
-            foreach (var versionDir in Directory.GetDirectories(packageDir))
+            foreach (var packageDir in Directory.GetDirectories(nugetCache))
             {
-                if (versionDir.EndsWith("-dev") || versionDir.EndsWith("-ci"))
+                foreach (var versionDir in Directory.GetDirectories(packageDir))
                 {
-                    Directory.Delete(versionDir, recursive: true);
+                    if (versionDir.EndsWith("-dev") || versionDir.EndsWith("-ci"))
+                    {
+                        Directory.Delete(versionDir, recursive: true);
+                    }
                 }
             }
         }
